+++ src/f_attractive_sector.c
@@ -43,11 +43,11 @@ attractive_sector_problem(const size_t number_of_variables,
 
   problem->problem_name = coco_strdup("attractive sector function");
   /* Construct a meaningful problem id */
-  problem_id_length = snprintf(NULL, 0, "%s_%02i", "attractive_sector",
-                               (int)number_of_variables);
+  problem_id_length = (size_t)snprintf(NULL, 0, "%s_%02lu", "attractive_sector",
+                               (long)number_of_variables);
   problem->problem_id = coco_allocate_memory(problem_id_length + 1);
-  snprintf(problem->problem_id, problem_id_length + 1, "%s_%02d",
-           "attractive_sector", (int)number_of_variables);
+  snprintf(problem->problem_id, problem_id_length + 1, "%s_%02lu",
+           "attractive_sector", (long)number_of_variables);
 
   problem->number_of_variables = number_of_variables;
   problem->number_of_objectives = 1;
