+++ src/f_lunacek_bi_rastrigin.c
@@ -119,10 +119,10 @@ lunacek_bi_rastrigin_problem(const size_t number_of_variables,
   problem->problem_name = coco_strdup("BBOB f24");
   /* Construct a meaningful problem id */
   problem_id_length =
-      snprintf(NULL, 0, "%s_%02i", "bbob2009_f24", (int)number_of_variables);
+      (size_t)snprintf(NULL, 0, "%s_%02lu", "bbob2009_f24", (long)number_of_variables);
   problem->problem_id = coco_allocate_memory(problem_id_length + 1);
-  snprintf(problem->problem_id, problem_id_length + 1, "%s_%02d",
-           "bbob2009_f24", (int)number_of_variables);
+  snprintf(problem->problem_id, problem_id_length + 1, "%s_%02lu",
+           "bbob2009_f24", (long)number_of_variables);
 
   problem->number_of_variables = number_of_variables;
   problem->number_of_objectives = 1;
